- refactored - show scrollbar when clip is moved
