+++ druid/src/widget/scroll.rs
@@ -197,10 +197,10 @@ impl<T: Data, W: Widget<T>> Widget<T> for Scroll<T, W> {
             if let Event::Notification(notification) = event {
                 if let Some(&global_highlight_rect) = notification.get(SCROLL_TO_VIEW) {
                     ctx.set_handled();
-
-                    self.clip.with_port(|port| {
-                        default_scroll_to_view_handling(ctx, port, global_highlight_rect);
-                    });
+                    let view_port_changed = self.clip.default_scroll_to_view_handling(ctx, global_highlight_rect);
+                    if view_port_changed {
+                        self.scroll_component.reset_scrollbar_fade(|duration|ctx.request_timer(duration), env);
+                    }
                 }
             }
         }
